refactor: UI 및 상태 관리 모듈 분해 리팩토링
대규모 리팩토링을 통해 UI와 상태 관리 로직을 담당하던 거대 모듈들을 기능에 따라 작고 독립적인 모듈로 분해했습니다. 이를 통해 코드의 복잡도를 낮추고 유지보수성과 가독성을 크게 향상시켰습니다.

주요 변경 사항:

- UI 모듈 분리: 거대했던 ui.js 파일을 cardRenderer.js, unitRenderer.js, stats.js로 분리하여 각 UI 요소의 렌더링 책임을 위임했습니다.

- 상태 관리 모듈 분리: state.js 파일의 여러 역할을 storage.js(저장소), dataLoader.js(데이터 로딩), migration.js(데이터 마이그레이션)로 분리했습니다.

- 버그 수정 및 효율화: 중복 코드가 많았던 modal.js의 로직을 통합하고, 창 크기 조절 시 발생하던 유닛 파괴 오버레이 버그를 디바운싱을 이용한 리렌더링으로 해결했습니다.
