+++ index.html
@@ -45,6 +45,8 @@
     </div>
 
     <div id="units-container"></div>
+    <div id="drones-container"></div>
+    <div id="tactical-cards-container"></div>
 
     <div class="add-button-container">
         <button id="add-unit-button">+ 유닛 추가</button>
