+++ modules/modal.js
@@ -1,27 +1,75 @@
 import * as dom from './dom.js';
 import * as state from './state.js';
-import { renderRoster, createBuilderModeImage } from './ui.js';
+import { renderRoster } from './ui.js';
+import { createBuilderModeImage } from './cardRenderer.js';
 import { performActionAndPreserveScroll } from './gameMode.js';
 import { CSS_CLASSES } from './constants.js';
 
-let currentUnitId = null;
-let currentCategory = null;
-let isBackCard = false;
+// --- Helper Functions ---
+
+const createCardItem = (cardData, clickHandler) => {
+    const cardItem = document.createElement('div');
+    cardItem.className = CSS_CLASSES.MODAL_CARD_ITEM;
+    const img = createBuilderModeImage(cardData);
+    cardItem.appendChild(img);
+
+    const points = document.createElement('div');
+    points.className = CSS_CLASSES.CARD_POINTS;
+    points.textContent = cardData.points || 0;
+    cardItem.appendChild(points);
+
+    cardItem.addEventListener('click', () => clickHandler(cardData));
+    return cardItem;
+};
+
+const createDeselectOption = (clickHandler) => {
+    const deselect = document.createElement('div');
+    deselect.className = CSS_CLASSES.DESELECT_OPTION;
+    deselect.textContent = '선택 해제';
+    deselect.addEventListener('click', clickHandler);
+    return deselect;
+};
+
+// --- Modal Core Logic ---
 
 export const closeModal = () => {
     dom.modalOverlay.style.display = 'none';
     dom.modalImageContainer.innerHTML = '';
-    dom.modalImageContainer.classList.remove(CSS_CLASSES.DRONE_VIEW);
+    // Reset container classes
+    dom.modalImageContainer.className = 'modal-image-container';
 };
 
-const addCardToUnit = (cardData) => {
+const _openModal = ({ title, cards, onCardClick, containerClasses = [], showDeselect = false, deselectHandler }) => {
+    dom.modalTitle.textContent = title;
+    dom.modalImageContainer.innerHTML = '';
+    dom.modalImageContainer.className = 'modal-image-container';
+    containerClasses.forEach(cls => dom.modalImageContainer.classList.add(cls));
+
+    const faction = state.getActiveRoster().faction;
+    const factionFilteredCards = cards.filter(c => c.faction === faction || c.faction === 'Public');
+
+    if (showDeselect && deselectHandler) {
+        dom.modalImageContainer.appendChild(createDeselectOption(deselectHandler));
+    }
+
+    factionFilteredCards.forEach(cardData => {
+        if (cardData.visible === false) return;
+        dom.modalImageContainer.appendChild(createCardItem(cardData, onCardClick));
+    });
+
+    dom.modalOverlay.style.display = 'flex';
+};
+
+// --- Card Action Handlers ---
+
+const createCardToUnitHandler = (unitId, category, isBack) => (cardData) => {
     performActionAndPreserveScroll(() => {
         const roster = state.getActiveRoster();
-        if (isBackCard) {
-            const drone = roster.drones.find(d => d.rosterId === currentUnitId);
+        if (isBack) {
+            const drone = roster.drones.find(d => d.rosterId === unitId);
             if (drone) drone.backCard = cardData;
         } else {
-            roster.units[currentUnitId][currentCategory] = cardData;
+            roster.units[unitId][category] = cardData;
         }
     });
 
@@ -47,76 +95,34 @@ const addTacticalCardToRoster = (cardData) => {
     closeModal();
 };
 
-const createCardItem = (cardData, clickHandler) => {
-    const cardItem = document.createElement('div');
-    cardItem.className = CSS_CLASSES.MODAL_CARD_ITEM;
-    const img = createBuilderModeImage(cardData);
-    cardItem.appendChild(img);
 
-    const points = document.createElement('div');
-    points.className = CSS_CLASSES.CARD_POINTS;
-    points.textContent = cardData.points || 0;
-    cardItem.appendChild(points);
-
-    cardItem.addEventListener('click', () => clickHandler(cardData));
-    return cardItem;
-};
-
-const createDeselectOption = () => {
-    const deselect = document.createElement('div');
-    deselect.className = CSS_CLASSES.DESELECT_OPTION;
-    deselect.textContent = '선택 해제';
-    deselect.addEventListener('click', () => addCardToUnit(null));
-    return deselect;
-};
+// --- Public-Facing Modal Openers ---
 
 export const openModal = (unitId, category, isBack = false) => {
-    currentUnitId = unitId;
-    currentCategory = category;
-    isBackCard = isBack;
-    dom.modalTitle.textContent = `${category} 부품 선택`;
-
-    const cardsToShow = state.allCards.byCategory[category] || [];
-    const faction = state.getActiveRoster().faction;
-    const factionFilteredCards = cardsToShow.filter(c => c.faction === faction || c.faction === 'Public');
-
-    dom.modalImageContainer.appendChild(createDeselectOption());
-    factionFilteredCards.forEach(cardData => {
-        if (cardData.visible === false) return;
-        dom.modalImageContainer.appendChild(createCardItem(cardData, addCardToUnit));
+    const clickHandler = createCardToUnitHandler(unitId, category, isBack);
+    _openModal({
+        title: `${category} 부품 선택`,
+        cards: state.allCards.byCategory[category] || [],
+        onCardClick: clickHandler,
+        showDeselect: true,
+        deselectHandler: () => clickHandler(null)
     });
-
-    dom.modalOverlay.style.display = 'flex';
 };
 
 export const openDroneModal = () => {
-    dom.modalTitle.textContent = '드론 선택';
-    dom.modalImageContainer.classList.add(CSS_CLASSES.DRONE_VIEW);
-
-    const faction = state.getActiveRoster().faction;
-    const factionFilteredDrones = state.allCards.drones.filter(d => d.faction === faction || d.faction === 'Public');
-
-    factionFilteredDrones.forEach(cardData => {
-        if (cardData.visible === false) return;
-        dom.modalImageContainer.appendChild(createCardItem(cardData, addDroneToRoster));
+    _openModal({
+        title: '드론 선택',
+        cards: state.allCards.drones,
+        onCardClick: addDroneToRoster,
+        containerClasses: [CSS_CLASSES.DRONE_VIEW]
     });
-
-    dom.modalOverlay.style.display = 'flex';
 };
 
 export const openTacticalCardModal = () => {
-    dom.modalTitle.textContent = '전술 카드 선택';
-    dom.modalImageContainer.classList.remove(CSS_CLASSES.DRONE_VIEW);
-    dom.modalImageContainer.classList.add(CSS_CLASSES.TACTICAL_CARD_VIEW);
-
-    const faction = state.getActiveRoster().faction;
-    const tacticalCards = state.allCards.tactical || [];
-    const factionFilteredTacticalCards = tacticalCards.filter(c => c.faction === faction || c.faction === 'Public');
-
-    factionFilteredTacticalCards.forEach(cardData => {
-        if (cardData.visible === false) return;
-        dom.modalImageContainer.appendChild(createCardItem(cardData, addTacticalCardToRoster)); // Tactical cards are added using addTacticalCardToRoster
+    _openModal({
+        title: '전술 카드 선택',
+        cards: state.allCards.tactical || [],
+        onCardClick: addTacticalCardToRoster,
+        containerClasses: [CSS_CLASSES.TACTICAL_CARD_VIEW]
     });
-
-    dom.modalOverlay.style.display = 'flex';
 };
