+++ style.css
@@ -51,6 +51,7 @@ h1 {
     flex-wrap: nowrap;
     overflow-x: auto;
     padding-bottom: 15px;
+    box-sizing: border-box; /* Prevent padding from affecting width calculations */
 }
 
 .card-wrapper {
@@ -287,7 +288,7 @@ h1 {
 
 .modal-image-container {
     display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
+    grid-template-columns: repeat(auto-fill, minmax(180px, 1fr));
     gap: 15px;
     overflow-y: auto;
     padding: 15px;
@@ -722,10 +723,11 @@ div.hidden-card-overlay > img.hidden-title-image {
     color: #594300;
 }
 
+/* Custom modal view styles */
 .modal-image-container.drone-view {
-    grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
+    grid-template-columns: repeat(auto-fill, minmax(360px, 1fr));
 }
 
 .modal-image-container.tactical-card-view {
-    grid-template-columns: repeat(auto-fill, minmax(239px, 1fr));
+    grid-template-columns: repeat(auto-fill, minmax(240px, 1fr));
 }
